+++ src/it/readdsmf.c
@@ -204,7 +204,7 @@ static int it_riff_dsmf_process_pattern( IT_PATTERN * pattern, DUMBFILE * f, int
 		++ row;
 	}
 
-	pattern->n_entries = entry - pattern->entry;
+	pattern->n_entries = (int)((long)entry - (long)pattern->entry);
 	if ( ! pattern->n_entries ) return -1;
 
 	return 0;
